refactor(worktree-ops): split worktreeOperations into git/state/tabCleanup/bridge/prompts submodules
Mirrors the precedent set by #114-#118: target file becomes a facade
that composes single-concern submodules. The 638-line file is now a
158-line facade plus 7 submodules under src/hooks/projectOps/worktreeOps/.

Public surface unchanged: useWorktreeOperations and the
WorktreeOperations / WorktreeOperationDeps types remain importable from
./worktreeOperations. The lone consumer (useProjectOps) is untouched.

The window.confirm / window.alert prompts are preserved as-is in
prompts.ts to keep this a pure refactor; a follow-up should swap them
to the project's notification + AskUserConfirm pattern.
