+++ src/hooks/projectOps/worktreeOps/lookups.ts
@@ -1,0 +1,28 @@
+import type { MutableRefObject } from "react";
+import type { Project, ProjectsState } from "../../../projects";
+import type { Worktree } from "../../../worktrees";
+import type { ProjectLookups } from "./types";
+
+export function makeProjectLookups(
+  projectsRef: MutableRefObject<ProjectsState>,
+): ProjectLookups {
+  function findProject(id: string): Project | null {
+    return projectsRef.current.projects.find((p) => p.id === id) ?? null;
+  }
+
+  function findProjectOfWorktree(
+    worktreeId: string,
+  ): { project: Project; worktree: Worktree } | null {
+    const wbp = projectsRef.current.worktreesByProject;
+    for (const [pid, list] of Object.entries(wbp)) {
+      const wt = list.find((w) => w.id === worktreeId);
+      if (wt) {
+        const project = findProject(pid);
+        if (project) return { project, worktree: wt };
+      }
+    }
+    return null;
+  }
+
+  return { findProject, findProjectOfWorktree };
+}
