+++ src/hooks/projectOps/worktreeOps/tabCleanup.ts
@@ -1,0 +1,66 @@
+import type { MutableRefObject } from "react";
+import type { Tab } from "../../../types/tab";
+import { normalizeSessionPath, projectScopeBucketKey } from "../tabBuckets";
+import type { TabBucket } from "../types";
+
+interface TabCleanupDeps {
+  stateRef: MutableRefObject<Record<string, unknown>>;
+  tabBucketsRef: MutableRefObject<Map<string, TabBucket>>;
+  syncRecentSessionsToState: () => void;
+  closeTabNow: (tabId: string) => void;
+  activateWorktree: (worktreeId: string | null) => void;
+}
+
+function tabCwdMatches(tab: Tab, path: string): boolean {
+  if (tab.kind !== "agent") return false;
+  return normalizeSessionPath(tab.cwd) === normalizeSessionPath(path);
+}
+
+function removeStoredTabsForWorktreePath(
+  tabBucketsRef: MutableRefObject<Map<string, TabBucket>>,
+  path: string,
+  removedBucketKey: string,
+): void {
+  tabBucketsRef.current.delete(removedBucketKey);
+  for (const [key, bucket] of tabBucketsRef.current.entries()) {
+    const tabs = bucket.tabs.filter((tab) => !tabCwdMatches(tab, path));
+    if (tabs.length === bucket.tabs.length) continue;
+    if (tabs.length === 0) {
+      tabBucketsRef.current.delete(key);
+      continue;
+    }
+    const activeTabId = tabs.some((tab) => tab.id === bucket.activeTabId)
+      ? bucket.activeTabId
+      : tabs[0]?.id;
+    tabBucketsRef.current.set(key, { tabs, activeTabId });
+  }
+}
+
+function closeVisibleTabsForWorktreePath(
+  stateRef: MutableRefObject<Record<string, unknown>>,
+  closeTabNow: (tabId: string) => void,
+  path: string,
+): void {
+  const tabs = (stateRef.current.tabs as Tab[] | undefined) ?? [];
+  const closing = tabs
+    .filter((tab) => tabCwdMatches(tab, path))
+    .map((tab) => tab.id);
+  for (const tabId of closing) closeTabNow(tabId);
+}
+
+export function closeTabsForRemovedWorktree(
+  deps: TabCleanupDeps,
+  projectId: string,
+  worktreeId: string,
+  path: string,
+  wasActive: boolean,
+): void {
+  closeVisibleTabsForWorktreePath(deps.stateRef, deps.closeTabNow, path);
+  if (wasActive) deps.activateWorktree(null);
+  removeStoredTabsForWorktreePath(
+    deps.tabBucketsRef,
+    path,
+    projectScopeBucketKey(projectId, worktreeId),
+  );
+  deps.syncRecentSessionsToState();
+}
